explored how R's quantile function works (at least in default case).
Edited fucntion "checkchao" to estimate the quantile of the true diversity in the bootstrapped asympttoic estimates slightly differently. Still not sure this is the right way to go, where right means that it matches the quantile function for the same distribution most closely. There might need to be some interpolation.

Explored whether the CI for coverage-based rarefaction were as bad as I suspected. yes.